small changes
read_image_list can read image files without filelist.txt
is_image_file checks if given path is an image file or not
small updates for noise generation functions

+++ code/generate_noisy_image.m
@@ -1,6 +1,6 @@
 function [ im_noisy ] = generate_noisy_image( im, sigma )
-noise_data = uint8(randn(size(im)) .* sigma);
-im_noisy = im + noise_data;
+noise_data = randn(size(im)) .* sigma;
+im_noisy = double(im) + noise_data;
 im_noisy(im_noisy < 0) = 0;
 im_noisy(im_noisy > 255) = 255;
 end
